feature: Add recommends and document package dependencies
This is to make it easier to get the package up and running. Users
who don't want to care about each individual dependency package can
simply install all the Recommends packages, and they should find that
everything can be used. Users who wish to only install the
packages they need can remove those packages which are Recommended
but not needed for their use case.

The package description is extended with information about the
optional features and dependencies they require, in order to help
the sysadmin choose which packages they need.

Closes LINOTP-1456