fix: Keep old apache realm name to avoid breakage when upgrading
As part of a number of changes to drop the '2' suffix, we also
modified the apache realm name. Unfortunately, this affects the
hashed admin password and prevents logging in on a system that
has been upgraded from version 2.

Here we go for the simplest fix possible: Revert the relevant changes
and continue to use the old name for now. We could make things a
little more clever, such as only keeping the suffix on upgraded
systems, but if we're going to spend time on this we'd rather spend
the time on a more modern replacement that is not apache specific.
In the meantime we'll live with the old name in this one place.

Closes LINOTP-1457