test: Work around problem loading debconf templates
debconf-set-selections does not like being asked to load questions
that it hasn't seen before. This means we can't answer new questions
just before upgrading.

In order to work around this, we change the preseeding step to run
the configuration script twice: The first time, we use
DEBIAN_FRONTEND=noninteractive to supply default answers to all the
new questions. Then we do the actual preseeding of the answers we
want to give.