test: Ensure errors during init commands are handled
If the tests were executed on a system which did not have an openssl
command available, test setup would fail with a confusing traceback
that complained about missing app context.

The problem could not be fixed by providing a real app context because
it is too early in the setup process.

So we fix this by providing a fake app context to the init commands
that just contains an echo command.