fix: Remove diagnostic message workaround
We had added two exception catchers to work around problems with
the init audit keys being called out of application context during
the tests. The problem was that we needed a try..except handler in
every echo() site, not just these two.

Now that this is fixed in the tests themselves we can
remove the exception catching.