fix: Regenerate database configuration when upgrading from 3.0 beta
We changed the database configuration setting name for 3.0rc0
(commit 6f30bedf). This needs the database configuration file
to be generated for this version, so we tell the infrastructure
to run a database upgrade for this version.

We do not need to run the upgrade twice when upgrading this package
from v2, so the 2.80 upgrade is no longer needed.

Part of LINOTP-1431