test: Fix use of configured database in tests
The configuration variable for database tests was changed in ab0ec9
to avoid accidentally overwriting a database. These fixes are needed
to make sure that the existing tests use the new variable name.

Note that as a convenience for debugging, we leave the
LINOTP_DATABASE_URI set in those environments that only contain a
throwaway database.