fix: Rename reporting columns back to 2.12 state.
This “fix” is in pursuit of LINOTP-1479, which claims a database error
when a token is enrolled while reporting is on. We've looked into this
and the problem could not be reproduced in a current Docker-based test
environment.

Even so it looks as if two columns in the `REPORTING` table had their
names changed during the Flask-SQLAlchemy migration. We're changing
these back on the off-chance, to avoid issues with older databases.

We're also committing a test for the erroneous behaviour just in case
this isn't a complete fluke after all.