fix: Add `SESSION_COOKIE_SECURE` setting and make LinOTP pay attention.
This adds the `SESSION_COOKIE_SECURE` setting (it defaults to
`True`). We also change `AdminController.getsession()` to obey this
setting so the session cookie has its `secure` flag set if
appropriate.

Note that `SESSION_COOKIE_SECURE` is really a Flask setting, and the
built-in Flask session machinery, which we're duly ignoring, uses it
automatically. We should really convert LinOTP to use Flask sessions
instead of the homegrown stuff, because (a) the homegrown stuff has a
certain … aroma, and (b) the Flask session handler does nice things
for us at no extra charge.