fix: [deb] import of linotp 2 settings on upgrade
The package configuration script can be run before or after the
preinst is executed, depending on the way the upgrade is run. [*]

This means that the rename of /etc/linotp2 -> linotp can happen
before or after the configure script is run. So we adjust the config
import to allow for this case.

Closes LINOTP-1486

[*] https://www.debian.org/doc/debian-policy/ch-binary.html#prompting-in-maintainer-scripts