test: integration test SelfService class refactoring
The SelfService class is now more independent from the URL path the
selfservice is run at. The URL is configured as a property in the class.
The SelfService class is following the style of the ManageUI class in
accepting the TestCase and providing properties for often accessed
parts of the testcase.