feature: change url_prefix of selfservice controller to /selfservice-legacy
This is done by introducing a mechanism that allows setting a url_prefix
default in the controller class itself. It will keep defaulting to
'/' + ctrl_name in the BaseController if the controller does not override
the default. It is still possible to set a different url_prefix on
configuration level by setting the CONTROLLERS config item.

Refers to LINOTP-1499