chore: allow changing the distribution independently from the debian version in makefile
The separate dependency apt file called kideps.list that is preconfigured
to install linotp dependencies was only allowing the distribution with the
same name as the debian release name used to select the debian docker
image.

It is now possible to set DEPENDENCY_{SORCE, DISTRBUTION, COMPONENT,
GPG_KEYID, GPG_KEY_URL} to fully configure the repository used for the
dependency repo.