fix: Don't attempt stealth STARTTLS for ldap:// connections w/o encryption.
In LinOTP 2.x, even `ldap://` connections with the STARTTLS option
explicitly set to “off” would attempt to do STARTTLS behind the
scenes, and only fall back to unencrypted LDAP if that failed. This is
too much magic for our taste, and by now, if you configure a new LDAP
resolver, the STARTTLS option defaults to “on”. We assume that if you
explicitly set it to “off”, you mean it. This was supposed to be
implemented in LINOTP-1385, but the omission somehow slipped through
the cracks.

The decision tree now looks like:

  - ldap:// URL
    - `enforce_tls` option set: do STARTTLS
    - `enforce_tls` option not set: don't do STARTTLS
  - ldaps:// URL
    - connection uses TLS from the get-go (no STARTTLS required)