fix: Adjust `linotp ldap-test` to cope with the new reality.
The difference here is that, for compatibility with LinOTP 2.x, in the
“ldap:// URL, STARTTLS, no valid CA certificate, no checking” case,
the test program used to expect the connection to fail because when
using LDAP+STARTTLS, LinOTP 2.x would check the CA certificate for
validity even if certificate checking was explicitly turned off by the
user. (If you wanted to sneak an invalid CA certificate past the LDAP
resolver, you would ask for an *unencrypted* LDAP connection, which
would try encryption anyway but would not worry about the CA
certificate. Great stuff.)

This is now fixed so the test program expects the connection to come
up if certificate checking is disabled, even if there is no valid CA
certificate available.