test: Add 'nodatabase' test to check installs without database
If linotp is installed by an installer such as the SVA, there will not
yet be a database service available. We can defer the configuration by
answering the debconf questions appropriately.

The previous workaround in LinOTP 2 was to create a temporary script
in rc.local that would create the database and then delete itself.

This tests confirms that this use case is properly supported by the
packaging scripts.