fix: Enrollment token description was not correctly set for validate
If an enrollment token is created without a description, it will
normally have the description 'rollout token', which makes it easier for
administrators to find and manage.

In the case that the enrollment token was created for usage in both the
self service and validate context, the description was not set correctly.

Related to LINOTP-1529