fix: Support hyphenated URLs with `@add_hyphenated_url`.
Our Pylons-compatibility scaffolding creates Flask URL routings
automatically based on the names of public methods in controller
classes. This means that a method like `auth.challenge_response` is
connected to the `/auth/challenge_response` URL.

This approach is OK in principle except that a few endpoints have been
documented with hyphens in them rather than underscores. This means we
should support `/auth/challenge-response` too, on the off-chance that
this is mentioned somewhere in older documentation that hasn't yet
been fixed.

In principle it would be very easy to add hyphenated paths as
alternatives to *all* paths with underscores, but we don't want to do
this because it would create ambiguity and possibly lead to further
maintenance issues down the road. Nor do we want to make a special
hard-coded exception for `/auth/challenge-response` because special
hard-coded exceptions are ugly.

Instead, we fix this by adding an `@add_hyphenated_url` decorator
whose presence on a controller-endpoint method will cause an
additional URL rule to be installed for that endpoint which has all
underscores replaced by hyphens. This is a reasonable compromise
because it avoids hard-coding endpoint URLs in the base controller and
confines the actual implementation details (an attribute on the
function object) to the `controller.base` module.