fix: update license via POST body
When we ported LinOTP to Python 3, we broke one of the ways to update the license.
There are two ways to do it: by sending it in the body of a POST request
or by uploading a file in a form.

The later was working, but the POST body one was not yet ported. We fix
that here now by checking the request parameters for a 'license' field.
Its value also does not need to be `decode()`d any longer.