chore: include tools script in setup.py
We used to include a lot of scripts in the setup.py's scripts declaration.
With the deprecation and removal of a lot of scripts, the scripts block
was completely removed from setup.py. We have one script remaining in the
tools folder that we hereby now again install as a script via setup.py.