feature: webprovision policies allow userinit method
The userinit method that lets a user create new tokens now accepts both
policies webprovisionGOOGLE and webprovisionGOOGLEtime to enroll HOTP
and TOTP tokens respectively.

This is possible since HOTP and TOTP tokens are now compatible with the
Google auth configuration if no configuration parameters (hashlib,
window, etc) are specified. The only thing that was missing was
accepting both policies, as if they were the enrollHMAC and enrollTOTP
policies.