feature: prevent ldap resolver test with unsaved resolver config
An admin without system write permission can execute the resolver test to test
the resolver configuration. This way, it was possible to invalidate the resolver
credentials by sending a wrong configuration to the ldap / ad resolver with only
read permission.

To prevent this abuse, we no longer allow testing unsaved resolver data and only
accept the "name" parameter to select the resolver to test. The UI must ensure that
the resolver is saved first before testing it.

Part of LINOTP-1524