fix: correctly reset ldap resolver input validation
The password field is a required field if, so called, "critical inputs" are changed.
If the password field is colored red with the text "password required" placeholder
and the edit is canceled, the field was not correctly reset for the next time,
the resolver edit dialog was opened. This is hereby fixed.