refactor: prevent form submission by changing button type
Before, clicking the AD/LDAP preset buttons was submitting the form.
This, however, usually triggers a page reload, which we were preventing
by interrupting it with a `return false` in the handling of the click
event.

Now we set the button types to button (overriding the default submit
type), so we achieve the same effect.