fix: wrong counts in token monitoring for mysql / sqlite
If tokens are assigned to multiple realms, the total token count was
off because those tokens were counted for each realm. This bug only
applies to sqlite and mysql. The fix is to not explicitly query all
tables because only postgres is allowing to specify a key to count
for DISTINCT.

The DISTINCT filtering was also applied to the token status filtering
where no measures where taken at all to filter tokens. These token
counts were counted twice even with postgres.

Closes LINOTP-1540