fix: strip and filter status param in token monitoring
This change does two things:

a) if no status param is sent or the status ends with a trailing comma, we no
longer include an empty field in each section of the Realms and Summary
dictionaries.

b) status values are stripped so that spaces next to a comma are no problem for
matching the search terms.

Detected and fixed during LINOTP-1540.