chore: refactor fallback dependency repo to restore Dockerfile behaviour
In commit 3c0a2e6f, some parts of the Dockerfile were removed because we
needed to quickly change the debian repository key retrieval. This commit
fixes the original problem that was bypassed with the removed code. It is now
possible to set DEPENDENCY_GPG_KEY_URL and unset DEPENDENCY_GPG_KEYID.