fix: audit database problems caused by not using flask_sqlalchemy
We now use flask_sqlalchemy to configure the audit database connection.
This fixes the problem where a separate mysql based database could not
be used as the audit database. We use a different bind for the audit
table to be able to use a different database for the audit connection.

The audit command line interface is now using a modern query style and
the flask_sqlalchemy database session.

Closes LINOTP-1596