fix: performance improvment for token finish
when the validation has been done, the resulting token lists like
pin_matching and invalid tokens are queried for pending challenges. When
there was a pin match of a token, but the otp was wrong, the same tokens are
in both lists and the challenges are queried twice for each token. Thus
putting the tokens into a set reduces these efforts so t only runs once.

Same holds true for the last modified time stamp setting which results
as well in an token update.