fix: validate token seed on otpkey field update
Co-authored-by: Mahyar Moghimi <mahyar.moghimi@netgo.de>

When a token first is created it is a more or less empty object. The
token update method fills in additional attributes such as the token
seed. It now verifies if the seed is correctly formatted for HMAC type
tokens, ultimately  preventing importing or generating tokens with
invalid seeds.

We also introduced a meaningful exception for when the seed is invalid.