feature: Actually enable JWT authentication.
This adds the `JWTMixin` to the `AdminController`, making the `login`
and `logout` methods accessible. The old `AdminController.logout()`
method is removed.

We're also arranging for the double-submit token to be sent with AJAX
requests in an HTTP header. This provides CSRF protection.

Together, these changes make it possible to run LinOTP's management
interface with JWT authentication. Yay.