test: adapt functional tests to jwt authentication
To enhance the performence of the authenthication during the functional tests,
the JWT authenthication got mocked.

To enable the mocking functionality for JWT the controller and action argument in
function make_authenticated_request must be passed as position argument,
which is now an iternal function. All test which have been using the
make_authenticated_request are adjusted to use now the appropriate method.