fix: jwt_exempt_methods recognizes hyphenated urls
methods decorated with `@add_hyphenated_url` were not properly processed
in jwt_check as exempted methods because the url was not known. This commit
changes the content of jwt_exempt_methods from url parts to the method name
inside the controller. The unittest `test_auth_challenge_response_hyphenated`
cought this.