feature: limit admin login to a specific realm
This commit adds a configureable admin realm name.  The config value is called
"ADMIN_REALM_NAME" and defaults to "LinOTP_admins". Only users that are part of
this realm are given access to the Manage-UI and other admin APIs.

For the tests:
* We extend the `admin-passwd` file with an
entry for the user `admin` that will be used in tests with the admin realm and
its resolvers.
* We create an admin flatfile resolver fixture that points to said file, and
add it to the admin realm.

Closes LINOTP-1662