test: reset username of test client after each request
It seems that several different request of the flask test client
run in the same app context. So also the the g object needs to be reset.
For several tests it is necessary to start the request with an empty
g.username property, so we now delete it manually after each request
with the scoped_authclient