fix: off-by-one error in `audit cleanup` cli
Adding more tests revealed that there is an off-by-one error.
In the tests, it can be shown that the number of remaining item count
in the database was off by one entry and the header row in the csv
file was not counted in for the tests.

This also implies that it should now be possible to set the same value for min and max.