docs: provide example config in /etc/linotp/linotp.cfg
Previously we generated an empty /etc/linotp/linotp.cfg file in the
postinst step of the Debian installation. Now we copy it from the LinOTP
repository directory config/etc (as specified in setup.py under
data_files).

This file now contains example config lines which can be uncommented to
override the default settings. It also explains the configuration
variables in the examples and refers the user to the `linotp config
explain` command which provides information on further configuration
variables.