fix: predefined JWT_SECRET_KEY should not be overwritten with encKey
When initializing the Flask app, we ensure that the JWT_SECRET_KEY is
set. Should the respective variable not yet exist or contain a value, we
set its value to the encription key specified by the SECRET_FILE
variable.

However, the code that loads this value was also being triggered when
the JWT_SECRET_VARIABLE was set, as we were mistaking a dictionary key
for an attribute (the Flask Config actually inherits from Python's
dictionary and can be treated as one according to the Flask
documentation - see [0]).

This issue was caught by newly created unit tests which have also been
added to this commit.

[0] https://flask.palletsprojects.com/en/2.0.x/config/