feature: add crud operations for local administrators
Library which is responsible for creating, reading, updating
and deleting local administrators in the local admin resolver.

In principle you can pass an unencrypted password to
`LocalAdminResolver.add_user()`, but in the CLI we prefer to make
password setting a separate operation, in order to be able to handle
password input specially. This means that when a new user account is
created, it should not be useable before a password is set.

We support this by setting the (encrypted) password to `*` when no
password (or an empty password) is passed to `add_user()`. The string
`*` can never result from the encryption of a valid password
candidate, and this effectively locks the account.