test: preserve admin realm and resolver in selenium cleanup
From now on, we can no longer drop all resolvers and realms in test cleanup
because we still need the admin realm to login in the next test.

The default realm now also needs to be set to the realm under test so that
the existing tests keep working because they expect to automatically
authenticate against the realm that is created for the test and not against the
admin realm.

To help with the realm matching, we now remove the resolver list from the realm name internally