fix: admin resolver name should be handled case sensitve
The resolver name should not be converted to lowercase.
Therefore the name of the default admin resolver is
"LinOTP_local_admins" as stated in the setting. Here we removed
all the occurences that resolver was converted to lower, in the
latest epic and correspondingly replaced the expectations of a
lower case default admin resolver in test cases.