test: fix broken selenium tests by removing files on the test runner
The tests failed because in "fill_element_from_dict" in linotp/tests/integration/linotp_selenium_helper/user_id_resolver.py:596 the file_filename input (that is a text input!!!) is not receiving the exact characters that we instruct selenium to fill into the form. This only happens if an existing filename is entered. If something else is put in there that is not an /etc/… filename, than it works. A quick fix was to delete the file on the tester as it is only used on the LinOTP container image.