chore: Add `python3-libsmpp` to `Suggests:`.
We want to call people's attention to the fact that the
`python3-libsmpp` package is a useful adjunct to LinOTP in some
situations. Specifically, you will likely need `python3-libsmpp` if
you plan to send SMS from LinOTP, but otherwise you can perfectly well
do without it.

(A quick reminder about Debian dependencies: `Depends: B` in package A
means that A won't work unless B is installed. `Recommends: B` means
that it would be very unusual to see A installed but not B.
`Suggests: B` means that B may be useful if you have A but is entirely
optional. The main practical difference is that by default, package
installation tools such as `apt` will install B during the
installation of A if it is “recommended” but not if it is
“suggested”. “Suggested” packages are never installed automatically.)