refactor: realm selection handling
We no longer rely on string parsing to identify the name and to
detect the default realm. Instead, the realm name and default status is
now written in a machine-readable way in the realm list dialog and
read on selection. This removes all the trouble with regex parsing of
the text representation.