feature: Add support for system ID/password to dummy SMPP server.
In the interest of being able to simulate a login failure during
tests, the dummy SMPP server now lets you specify a system ID (what
reasonable people would otherwise call a “user name”) and
password. The system ID defaults to `smsclient` but is only ever
looked at if a non-`None` password was specified. If a password is
given during server initialisation and it is not `None`, then the
system ID and password values submitted as part of the
`bind_transceiver` operation must match the ones in the server, or an
`ESME_RBINDFAIL` error will be returned.