refactor: Revisit SMPP server initialisation.
We're adding a `DummySMPPServer.run_server()` method which
incorporates all socket operations. This method can be passed to
`threading.Thread()` and keeps the socket stuff inside the thread,
which avoids nasty issues.

This means that the initial socket setup must move from the
`DummySMPPServer()` constructor to `DummySMPPServer.__enter__()` for
the context-manager case. This doesn't make a big difference.