chore: Add `python3-smpplib` to the Dockerfile, just to be safe.
The Debian `control` file only `Suggests: python3-smpplib`, but we
need it for automated tests; also it is a hassle to add it to the
Docker image after the fact.

Note that we have to add another `apt-get update` after we add the
LinOTP Debian repository, to ensure that the packages from there are
actually picked up by `apt-get install`.