fix: for localisation within linotp cli commands
the 'linotp support' command has to deal with various license problems
whereby the responses are user visable and thus are localized. But the
command interfaces does not have a request context per se. In this
case the localistation is using the babel default_locale as fallback
which is in the case of linotp available langueges 'en'. This is a good choice
for the linotp cli as these responses are parsed eg in the scope of the
appliance.