fix: change log level for violates the DRY principle
The get method for flask config items get overwritten. The config
dictonary contains all configuration items.
The critical ones are thos items which contain a _file or _dir
in the key. So just for this keys a warning is needed if the
default value is not set corrcectly.
In all other cases the debug level is enough.