feature: If JWT_SECRET_KEY is not given, use PBKDF2 on encKey.
If no explicit JWT_SECRET_KEY is given in the configuration, then
instead of using the first encKey as the JWT_SECRET_KEY directly, run
it through PBKDF2 first.

This brings us two new configuration settings, namely
JWT_SECRET_SALT (we wouldn't want to use the same salt across all
LinOTP installations, would we now?) and
JWT_SECRET_ITERATIONS (because we can, and because we don't like
hard-coded magic numbers in our code).

Leaving JWT_SECRET_SALT empty creates a new random salt whenever
LinOTP starts, but that would mean that any outstanding JWTs no longer
work, which might be an issue, so this is not the default case. This
means that we do end up with a hard-coded *default* salt after all,
which security-minded folks may want to change.