refactor: Refactor a weird part of `policy.checkAdminAuthorization()`.
If we keep the outer `if user.realm: …`, the final comparison against
`user.resolver_config_identifier` is never reached because the whole
block presupposes that `user.realm` is truthy, therefore the `if
user.realm != "":` always succeeds and leads to a `return`.

Of the first two `if` expressions inside the block, if the first one
succeeds then the second one would also succeed, and since the
dependent block is exactly the same in both cases we can get rid of
the first `if` altogether.