fix: Send the csrf token for download forms
The download forms were not sending the csrf token and therefore
always respond with an unauthorized response.

The requests which led to downlowding a file, including:
Export Token Info
Export Audit Trail
Where not functioning with jwt authentication and therefore
always respond with an unauthorized response.
The reason was that the csrf token was not sent in the header
of the request.
So now we are sending the csrf token in the form.