Revert "Merge branch 'feature/user_and_resolver_caching_by_default' into 'master'"
After the analysis of caching behavior we noticed that in some cases
changes to the information of a user in a resolver (user being deleted
from one resolver or moving from one resolver to another) will not
automatically reflect in LinOTP because of this caching. Untill then
that this is fixed, we are gonna keep caching as an option.

LINOTP-1721

This reverts commit 1458ca3f41b066f16490bdf1e8842c0d8a5e3e34, reversing
changes made to c66f875aba35ee67b0827740f325bfd479002138.