chore: adding comments for jwt secret key config
The JWT_SECRET_KEY config, will set a constant key for the server.
This means that after a system restart the blocklist (in memory) would
be erased. Therefore the revoked tokens will still be valid unless the
new server instance is using a new key.
Since we already have it implemented, we set the JWT_SETCRET_KEY to
None so that every new instance of the linotp server would have
different secret key to produce the jwt tokens.