feature: remove JWT_SECRET_KEY from config
We want to make sure that every run of the server would have a
different JWT_SECRET_KEY so that the tokens produced by one instance
would not be valid for the next. This is a work around for preventing
revoked tokens (after a log out) to be used again after a server
restart. The reason is that the blocklist is implemented in the memory
and hence restarting the server with the same secret key will make
them valid.
Also, so far, only the LinOTP itself needs to know thejwt_secret_key.