feature: safeguard controller configuration for 'gettoken'
to ensure that it takes at least one administrative action on a linotp server
in addition to the linotp policy defintion to enable the 'gettoken'
controller.
In the scope of this ticket two new configuration values are defined:
DISABLE_CONTROLLERS and ENABLE_CONTROLLERS replace the CONTROLLERS
definition. This will ease the dedicated enabling or disabling of
controllers.