feature: in the policy processing we use the _getAuthenticated user
which returns the autheticated user as User class object. Formerly a
dictionary was used for the authenticated user. Thus all subscriptions
to the authenticated user eg. user.get['login'] has to be adjusted.
As the policy processing is able to deal with user objects, the identified
user object could be forwarde to the folicy evaluation and no interface has
to be changed.