fix: check that the resolver of the user is the right one
in the policy evaluation the get_uid_resolver() is used to identify the
user and to compare him with the policy conditions. With introducing the
authenticated user as object, the resolver specification in the
constructure is used to preserve the outcome of the user identification.
This resolverspec though has not been check during the call of
get_uid_resolver(), which would have returned all resolvers / the first
resolver which had a user of this name. To prevent this we now verify
that if a resolver spec is set, that the resolver, where the user is
checked for existance is only that one.

Background: it is possible to have multiple admins with same name in the
admin realm, which belong to different resolvers - during identification
of the user, the password is checked against all resolvers and the
matching one is preserved in the jwt identity.

In the admin policies it is required to distinguish between the
resolver to prevent a privilege escalation.