feature: preserve username, resolver and realm of a jwt authenticated user
before linotp 3.2 the admin authentication was done on the apache
authentication and the provided login name was sufficient in the admin
authentication.

with jwt based authetication the whole user could now be identified with
its loginname and resolver and realm. Thus now the admin policies could
distinguish between the users of different resolvers, which is a
prerequisit to prevent an admin priviledge escalation.