fix: cherry-pick of changes in logging for branch v3.2
The changes were first applied to the master in order to add timestamp
to the log. Also these changes would add possibility of formatting the
two different logs, namely console and file, seperately.
The naming conventino has also changed as described below:

refactor: Coherent naming for logging settings

LOG_FILE should of course refere to logging done in the file.
To be consistent we added the "_" to have both LOG_FILE and
LOG_CONSOLE.
I also changed the sqlalchemy_log_X to log_sqlalchemy_x to be
consistent with the rest of the log* settings.