refactor: nolonger rely on page reloads for login redirects
This helps us seperating the concerns of the backend providing the static html
and javascript and the frontend of directly moving to the desired view instead
of knowing when to ask the backend to figure out where to move the user to.

The ng manage ui will also do it this way.

We remove the necessity to have a valid user authenticated for /mange/ and
/manage/login to check whether to redirect the user to the login view or not.