feature: do not show jquery-migrate warnings in the browser console in debug mode
The jquery-migrate warnings are there to show you were to look for missing
migrations during development. Therefore the messages are not shown for
the minimized versions and this change is only required in two places.

We do ingore those warnings because we have no intention of migration
jquery to the latest version. Wie use jquery-migrate only as a compat layer
and want to ignore any message.