fix: close the welcome screen and menues before log out
In case of log out, it might be the case that some menues are open or
the welcome screen is still on the view, which can prevent the log out
to function correctly (because it is based on clicking on an
item). Hence we make sure that they are all closed.