fix: state modifying api endpoints are only accessible via POST requests
The background is to prevent XCRSF access from other pages in a browser
as the browser would provide the stored access cookies when accessing
this site. To prevent the access of a js script from a different browser
tab, the js script in that tab must know the uri and the XCRF token from
the cookie and provide this in the header of the request. As the js in
the other tab has no direct access to the cookies, no harmfull request
could be submitted after applying this commit.