fix: prevent trying to drop security module in cases of static requests
We were not loading the security module (hsm) for static modules, but
at the time of finalizing the request we were trying to drop
it. Although this was not a security concern, it was producing
unnecessary debug info for all static requests stating that the
security module was not found. With this commit we prevent that.