fix: The base controller global request_param
The BaseController.request_param was holding the value of the request
parameters which could be changing with the next request. This was
contaminating the BaseController object (which might be reused for
next requests) with global and changing variables.
In this comit we make it dynamic so it can be read fresh from the app
object everytime it is needed.