fix: fix for flaky audit key error
the audit public & private key is generated with openssl by the linotp
command line interface. The audit private key is used for the signing of
the audit entries. The signing is done by the pycryptodomex library,
which has a problem in identifying the prime number which was used by
openssl to generate the audit keys. The problem was solved with newer
versions of pycryptodomex. Thus the version pinning of pycryptodomex lib
could be removed.