fix: index out of bound error during restore
in case that in between the backup and restore a config entry or a token
was deleted, the query for the update will return no result, which will
result in an index out of bound error.
To enable better tracing the processing the config name and the token
serial is written to stdout.