refactor: remove useless check
The first permission check of system/getRealms happens in the __before__
method. It fails if the user does not have read permissions in system.
That automatically allows the user to view all the realms in the system.

Verifying admin rights for a particular realm is a useless check that
will never restrict the returned list, and we can remove this part of
the code.