test: add test for the /api/v2/tokens api
which cover the permission testing, the verification of the access and
the tokenlist respone with pagination.
we verify that the /api/v2/token could only be access with the given rights
which enforces that the admin must be allowed via admin/show policy to
view tokens. If not given an http 403 error will be raised.