feature: samlcheck to return admin attribute
For the linotp identity provider to know if a user is admin or not,
the samlcheck endpoint should return an attribute indicating this.
Here check if the user is part of the linotp local admin realm but
in future implementations we need to include the possibility of other
realms being used for administrator users.