feature: return all token dates in ISO 8601 format
The creation date, last successful authentication and last
authentication match for a token were being returned as provided by the
token iterator: a string formatted according to LinOTP's
"DEFAULT_TIMEFORMAT", specified in linotp/lib/utils.py

It looked like the timezone was being set to UTC by formatting the time
as a string terminated in GMT, however the datetime object was unaware
of the timezone. Internally we store all the dates in UTC but without
the tzinfo ("naive datetime"). In order to later not get the timezone
wrong, we now force the datetime to know it is in the UTC timezone
("aware datetime") in the Tokens controller only.

Therefore the Tokens controller now returns all date strings as provided
by `datetime.datetime.isoformat()` in the timezone
`datetime.timezone.utc`.