refactor: split getTokens4UserOrSerial by return type
There are now two ways to retrieve a list of tokens: one returns a list
of Token objects, and the other a list of TokenClass objects.

This is an important distinction, as they have a completely different
API.

When developing it is now clear what type of "token" object we are
dealing with, saving us some headaches.