refactor: rename variables and add types to createTokenClassObject
The class registry returns a TokenClass constructor when given a type.
Because it is a generic class registry, it is also used for other types
such as Resolvers and we cannot define the type annotation there.

Therefore we can only define the TokenClass type here. This solves the
issue that the return type of createTokenClassObject was being a union
of TokenClass and Any. Now it's cleary always a TokenClass.