feature: create resolver-related classes for use in the new API
In order to better work within the new API endpoints we define new
classes that act as an adapter by converting the fairly opaque
dictionaries into clearly typed objects.

The ResolverType class is an enum of resolver types. It maps to the
existing type strings in the resolver configuration definitions.

The User is a data class that represents the users returned by
resolver queries.

The Resolver represents an instance of a resolver. Currently the list of
resolvers is loaded from the LinOTP Config object, which is not very
easy to work with.

In the long run we would like to have a table for the resolvers in the
database, hence we place the Resolver class here in the `model` module.
Someday it will hopefully inherit from a SQLAlchemy Schema class... For
now it remains as an OOP wrapper for the dictionaries we drag around the
codebase.