feature: return simplified resolver info of token owner
When retrieving a token with the new API, the token owner fields for
resolver name and resolver class where being filled with the very
non-informative dot-separated resolver spec strings. Now the name
returns only the user-assigned name of the resolver, and the class is
just a string of the resolver type.