feature: offline support with forward token
when offline support for the qr token is required and the qr token is
proxied by a forward token, the forward token will have to ask for the
offline info of the target token.

to support the dynamic request, if the target token supports offline,
the 'supports_offline_mode' has become a property, which queries the
target token for this capability.