refactor: remove meaningless check_session method from controllers
check_session used to check the authorization. While removing the
authorization (because we now use flask-jwt-extended to do it),
I missed the fact that the remaining part is useless now.

I thought, it would do the client check, to verify whether a client ip
address is allowed at all before checking the session. Look at 1cff3fa70761a63a11e85393c2c7487c6c3c9b85
to see how comedic the commit is now with the new knowledge.

We do not have a replacement for `config.get("linotpNoSessionCheck")` and we
do not plan to add the feature back for now.

As this check was the last part of the function, all calls and the function
itself are hereby removed.